Extract Stock photos Group and Media Item to separate files
