+++ WordPress/Classes/ViewRelated/Aztec/Stock Photos/StockPhotosDataSource.swift
@@ -1,69 +1,5 @@
 import WPMediaPicker
 
-final class StockPhotosMediaGroup: NSObject, WPMediaGroup {
-    func name() -> String {
-        return "group"
-    }
-
-    func image(with size: CGSize, completionHandler: @escaping WPMediaImageBlock) -> WPMediaRequestID {
-        return 0
-    }
-
-    func cancelImageRequest(_ requestID: WPMediaRequestID) {
-        //
-    }
-
-    func baseGroup() -> Any {
-        return ""
-    }
-
-    func identifier() -> String {
-        return "group id"
-    }
-
-    func numberOfAssets(of mediaType: WPMediaType, completionHandler: WPMediaCountBlock? = nil) -> Int {
-        return 10
-    }
-}
-
-final class StockPhotosMedia: NSObject, WPMediaAsset {
-    func image(with size: CGSize, completionHandler: @escaping WPMediaImageBlock) -> WPMediaRequestID {
-        return 0
-    }
-
-    func cancelImageRequest(_ requestID: WPMediaRequestID) {
-        //
-    }
-
-    func videoAsset(completionHandler: @escaping WPMediaAssetBlock) -> WPMediaRequestID {
-        return 0
-    }
-
-    func assetType() -> WPMediaType {
-        return .image
-    }
-
-    func duration() -> TimeInterval {
-        return 0
-    }
-
-    func baseAsset() -> Any {
-        return ""
-    }
-
-    func identifier() -> String {
-        return "image"
-    }
-
-    func date() -> Date {
-        return Date()
-    }
-
-    func pixelSize() -> CGSize {
-        return .zero
-    }
-}
-
 final class StockPhotosDataSource: NSObject, WPMediaCollectionDataSource {
     func numberOfGroups() -> Int {
         return 1
